requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
pandas>=1.3.0
openpyxl>=3.0.7
python-docx>=0.8.11
pymysql>=1.0.2
cryptography>=3.4.7
python-dotenv>=0.19.0